python/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy[asyncio]
asyncpg
pyjwt
python-multipart
